refactor: Register Claude Code as explicit provider plugin, remove defaults
- Create claude.ts provider plugin with classify, buildInterruptContent,
  and buildControlResponse methods
- Move Claude Code classification logic from messageClassification.ts
  into the Claude Code plugin
- classifyMessage now always dispatches through plugin registry — no
  inline fallback for any provider
- handleInterrupt and handleControlResponse dispatch through plugin
  with error logging for unsupported providers — no implicit defaults
- messageClassification.ts reduced to thin dispatcher + CSS helpers
