+++ frontend/src/components/chat/messageClassification.ts
@@ -22,158 +22,22 @@ export type MessageCategory
     | { kind: 'compact_summary' }
     | { kind: 'unknown' }
 
-function isObj(v: unknown): v is Record<string, unknown> {
-  return typeof v === 'object' && v !== null && !Array.isArray(v)
-}
-
-/** Check whether the wrapper envelope represents a notification thread. */
-function isNotificationThreadWrapper(wrapper: { messages: unknown[] } | null): wrapper is { messages: unknown[] } {
-  if (!wrapper || wrapper.messages.length < 1)
-    return false
-  const first = wrapper.messages[0] as Record<string, unknown>
-  const t = first.type as string | undefined
-  const st = first.subtype as string | undefined
-  return t === 'settings_changed' || t === 'context_cleared' || t === 'interrupted' || t === 'rate_limit' || t === 'plan_execution' || t === 'agent_renamed'
-    || (t === 'system' && st !== 'init' && st !== 'task_notification')
-}
-
 /**
  * Classify a parsed message into exactly one category.
  *
- * This replaces the ~15 boolean flags previously computed in MessageBubble
- * and the repeated type-checking done by each renderer in the chain.
+ * Always dispatches through the provider plugin registry. Each provider
+ * (Claude Code, Codex, etc.) registers its own classify implementation.
  */
 export function classifyMessage(
   parentObject: Record<string, unknown> | undefined,
   wrapper: { old_seqs: number[], messages: unknown[] } | null,
   agentProvider?: AgentProvider,
 ): MessageCategory {
-  // Delegate to provider plugin if available and not the default Claude Code.
-  if (agentProvider != null
-    && agentProvider !== AgentProvider.UNSPECIFIED
-    && agentProvider !== AgentProvider.CLAUDE_CODE) {
-    const plugin = getProviderPlugin(agentProvider)
-    if (plugin)
-      return plugin.classify(parentObject, wrapper)
-  }
-  // 0. Empty wrapper (all notifications consolidated to no-ops) — hide.
-  if (wrapper && wrapper.messages.length === 0)
-    return { kind: 'hidden' }
-
-  // 1. Notification thread (wrapper with notification-type first message)
-  if (isNotificationThreadWrapper(wrapper)) {
-    // Filter out "allowed" rate limit messages — they are informational only
-    // and the popover already shows them.
-    const msgs = wrapper.messages.filter((m) => {
-      if (!isObj(m))
-        return true
-      return !(m.type === 'rate_limit' && isObj(m.rate_limit_info)
-        && (m.rate_limit_info as Record<string, unknown>).status === 'allowed')
-    })
-    if (msgs.length === 0)
-      return { kind: 'hidden' }
-    return { kind: 'notification_thread', messages: msgs }
-  }
-
-  if (!parentObject)
-    return { kind: 'unknown' }
-
-  const type = parentObject.type as string | undefined
-  const subtype = parentObject.subtype as string | undefined
-
-  // 2. Hidden: system init, or system status (non-compacting)
-  if (type === 'system') {
-    if (subtype === 'init')
-      return { kind: 'hidden' }
-    if (subtype === 'status' && parentObject.status !== 'compacting')
-      return { kind: 'hidden' }
-    // 3. Task notification
-    if (subtype === 'task_notification')
-      return { kind: 'task_notification' }
-    // 4. Other system subtypes are notifications (compact_boundary, microcompact_boundary, etc.)
-    return { kind: 'notification' }
-  }
-
-  // 4b. Non-system notification types
-  if (type === 'rate_limit') {
-    if (isObj(parentObject.rate_limit_info) && (parentObject.rate_limit_info as Record<string, unknown>).status === 'allowed') {
-      return { kind: 'hidden' }
-    }
-    return { kind: 'notification' }
-  }
-  if (type === 'interrupted' || type === 'context_cleared' || type === 'settings_changed' || type === 'agent_renamed')
-    return { kind: 'notification' }
-
-  // 5. Result divider
-  if (type === 'result')
-    return { kind: 'result_divider' }
-
-  // 6. Compact summary
-  if (parentObject.isCompactSummary === true)
-    return { kind: 'compact_summary' }
-
-  // 7. Control response (synthetic message with controlResponse)
-  if (parentObject.isSynthetic === true && isObj(parentObject.controlResponse))
-    return { kind: 'control_response' }
-
-  // 8. Assistant messages
-  if (type === 'assistant') {
-    const message = parentObject.message as Record<string, unknown> | undefined
-    if (isObj(message)) {
-      const content = (message as Record<string, unknown>).content
-      if (Array.isArray(content)) {
-        const contentArr = content as Array<Record<string, unknown>>
-        // Check for tool_use first (higher priority than text)
-        const toolUse = contentArr.find(c => isObj(c) && c.type === 'tool_use') as Record<string, unknown> | undefined
-        if (toolUse) {
-          return {
-            kind: 'tool_use',
-            toolName: String(toolUse.name || ''),
-            toolUse,
-            content: contentArr,
-          }
-        }
-        // Check for text content
-        const hasText = contentArr.some(c => isObj(c) && c.type === 'text')
-        if (hasText)
-          return { kind: 'assistant_text' }
-        // Check for thinking content
-        const hasThinking = contentArr.some(c => isObj(c) && c.type === 'thinking')
-        if (hasThinking)
-          return { kind: 'assistant_thinking' }
-      }
-    }
-    return { kind: 'unknown' }
-  }
-
-  // 9–10. User messages
-  if (type === 'user') {
-    const message = parentObject.message as Record<string, unknown> | undefined
-    if (isObj(message)) {
-      const content = (message as Record<string, unknown>).content
-      // 9. String content → user_text
-      if (typeof content === 'string')
-        return { kind: 'user_text' }
-      // 10. Array content with tool_result → tool_result
-      if (Array.isArray(content)) {
-        const hasToolResult = (content as Array<Record<string, unknown>>).some(
-          c => isObj(c) && c.type === 'tool_result',
-        )
-        if (hasToolResult)
-          return { kind: 'tool_result' }
-      }
-    }
-    return { kind: 'unknown' }
-  }
-
-  // 11. Plain object with string .content and no .type → user_content (hidden if flagged)
-  if (!type && typeof parentObject.content === 'string') {
-    if (parentObject.hidden === true)
-      return { kind: 'hidden' }
-    return { kind: 'user_content' }
-  }
-
-  // 12. Fallback
+  const provider = agentProvider ?? AgentProvider.CLAUDE_CODE
+  const plugin = getProviderPlugin(provider)
+    ?? getProviderPlugin(AgentProvider.CLAUDE_CODE)
+  if (plugin)
+    return plugin.classify(parentObject, wrapper)
   return { kind: 'unknown' }
 }
