+++ frontend/src/components/chat/providers/index.ts
@@ -1,4 +1,5 @@
 // Import all provider modules to trigger side-effect registrations.
+import './claude'
 import './codex'
 import './stubs'
